v2v: change the tmp dir to  for host_no_space case

+++ v2v/tests/src/specific_kvm.py
@@ -433,20 +433,25 @@ def run(test, params, env):
         """
         Create a large file to make left space of root less than $left_space MB
         """
-        cmd_df = "df -m / --output=avail"
+        cmd_guestfish = "guestfish get-cachedir"
+        tmp_dir = session.cmd_output(cmd_guestfish).split()[-1]
+        logging.debug('Command output of tmp_dir: %s', tmp_dir)
+        cmd_df = "df -m %s --output=avail" % tmp_dir
         df_output = session.cmd(cmd_df).strip()
         logging.debug('Command output: %s', df_output)
         avail = int(df_output.strip().split('\n')[-1])
         logging.info('Available space: %dM' % avail)
-        if avail > left_space - 1:
-            tmp_dir = data_dir.get_tmp_dir()
-            if session.cmd_status('ls %s' % tmp_dir) != 0:
-                session.cmd('mkdir %s' % tmp_dir)
-            large_file = os.path.join(tmp_dir, 'file.large')
-            cmd_create = 'dd if=/dev/zero of=%s bs=1M count=%d' % \
-                         (large_file, avail - left_space + 2)
-            session.cmd(cmd_create, timeout=v2v_timeout)
-        logging.info('Available space: %sM' % session.cmd(cmd_df).strip())
+        if avail <= left_space - 1:
+            return None
+        if not os.path.exists(tmp_dir):
+            os.mkdir(tmp_dir)
+        large_file = os.path.join(tmp_dir, 'file.large')
+        cmd_create = 'dd if=/dev/zero of=%s bs=1M count=%d' % \
+                     (large_file, avail - left_space + 2)
+        session.cmd(cmd_create, timeout=v2v_timeout)
+        newAvail = int(session.cmd(cmd_df).strip().split('\n')[-1])
+        logging.info('New Available space: %sM' % newAvail)
+        return large_file
 
     @vm_shell
     def corrupt_rpmdb(**kwargs):
@@ -781,7 +786,7 @@ def run(test, params, env):
             check_boot()
         if checkpoint.startswith('host_no_space'):
             session = aexpect.ShellSession('sh')
-            create_large_file(session, 1000)
+            large_file = create_large_file(session, 1000)
             if checkpoint == 'host_no_space_setcache':
                 logging.info('Set LIBGUESTFS_CACHEDIR=/home')
                 os.environ['LIBGUESTFS_CACHEDIR'] = '/home'
@@ -930,8 +935,7 @@ def run(test, params, env):
                 else:
                     service_mgr.stop('firewalld')
         if checkpoint.startswith('host_no_space'):
-            large_file = os.path.join(data_dir.get_tmp_dir(), 'file.large')
-            if os.path.isfile(large_file):
+            if large_file and os.path.isfile(large_file):
                 os.remove(large_file)
         # Cleanup constant files
         utils_v2v.cleanup_constant_files(params)
